refactor(company): trigger deal recalculation from the writing repository
Six of the seven recalculation triggers already sat in the repository
performing the write. The company settings path was the exception: it
declared recalculateDealWeightedValues on the repo port and had the
interactor call it under a condition.

That exception existed because the company repository wrote stage
weights straight into CustomColumn.options with its own prisma call,
bypassing the repository that owns custom columns and already carries
the trigger. The owning repository never saw the write, so the
interactor had to compensate.

Stage weights now go through PrismaCustomColumnRepo.setOptionWeights,
which owns the options shape and recalculates when the column drives
the forecast. Designating a column recalculates from updateDetails.
The port method and the interactor call are gone, so no interactor
triggers recalculation and one rule holds everywhere: the repository
performing a write restores the derived values it invalidates.

Recalculation stays cheap to repeat because recalculateTotals skips
the update when the stored values already match.
