+++ features/company/prisma-company.repository.ts
@@ -15,7 +15,7 @@ import type { CreateAuthLinkSubscriptionRepo } from "@/ee/messaging/connect/crea
 
 import { SubscriptionStatus } from "@/generated/prisma";
 
-import { getDealRepo } from "@/core/di";
+import { getCustomColumnRepo, getDealRepo } from "@/core/di";
 import { BypassTenantGuard } from "@/core/decorators/bypass-tenant.decorator";
 import { Transaction } from "@/core/decorators/transaction.decorator";
 import { BaseRepository } from "@/core/base/base-repository";
@@ -44,6 +44,8 @@ export class PrismaCompanyRepo
       data: { ...args, id: companyId },
       where: { id: companyId },
     });
+
+    if (args.dealWeightingColumnId !== undefined) await getDealRepo().recalculateWeightedValuesForCompany();
   }
 
   async getDetails() {
@@ -72,28 +74,7 @@ export class PrismaCompanyRepo
 
     if (!company?.dealWeightingColumnId) return;
 
-    const column = await this.prisma.customColumn.findFirst({
-      where: { id: company.dealWeightingColumnId, companyId },
-      select: { id: true, options: true },
-    });
-
-    const stored = (column?.options as { options?: unknown } | null)?.options;
-
-    if (!column || !Array.isArray(stored)) return;
-
-    const weightByOptionValue = new Map(entries.map((entry) => [entry.optionValue, entry.weight]));
-
-    const options = (stored as Array<Record<string, unknown>>).map((option) =>
-      typeof option.value === "string" && weightByOptionValue.has(option.value)
-        ? { ...option, weight: weightByOptionValue.get(option.value) }
-        : option,
-    );
-
-    await this.prisma.customColumn.update({ where: { id: column.id, companyId }, data: { options: { options } } });
-  }
-
-  async recalculateDealWeightedValues() {
-    await getDealRepo().recalculateWeightedValuesForCompany();
+    await getCustomColumnRepo().setOptionWeights(company.dealWeightingColumnId, entries);
   }
 
   @Transaction
